Add code for seabird tracking database maps
The raw data needed to run this code and the results are not in the repo due to data use restrictions. The data was requested from data owners from the BirdLife International Seabird Tracking Database (www.seabirdtracking.org). Final csv and plots of distributions in c-square grid per month per species are in the WKBB sharepoint (no raw data).